Move ChangedJobConfigurationProcessor
